pyproject.toml
[tool.poetry]
name = "animalpicsandvideos"
version = "0.1.0"
description = "See videos and pictures of animals"
authors = ["Pricilla Patricia de Aragao <pricillapatricia@gmail.com>"]

[tool.poetry.dependencies]
python = "3.8.3"
flask = "^1.1.2"
gunicorn = "^20.0.4"
requests = "^2.24.0"
psycopg2-binary = "^2.8.6"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
pre-commit = "^2.7.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
